refactor: move upload logic to separate module
